+++ backend/libraries/types/src/notifications.rs
@@ -68,6 +68,8 @@ pub struct BotChatEvent {
     pub event_index: EventIndex,
     #[serde(rename = "l")]
     pub latest_event_index: EventIndex,
+    #[serde(rename = "u")]
+    pub initiated_by: Option<UserId>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
@@ -80,10 +82,14 @@ pub struct BotCommunityEvent {
     pub event_index: EventIndex,
     #[serde(rename = "l")]
     pub latest_event_index: EventIndex,
+    #[serde(rename = "u")]
+    pub initiated_by: Option<UserId>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub enum BotLifecycleEvent {
+    #[serde(rename = "r")]
+    Registered(BotRegisteredEvent),
     #[serde(rename = "i")]
     Installed(BotInstalledEvent),
     #[serde(rename = "u")]
@@ -110,6 +116,14 @@ pub struct BotUninstalledEvent {
     pub uninstalled_by: UserId,
     #[serde(rename = "l")]
     pub location: BotInstallationLocation,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct BotRegisteredEvent {
+    #[serde(rename = "i")]
+    pub bot_id: UserId,
+    #[serde(rename = "n")]
+    pub bot_name: String,
 }
 
 impl Debug for UserNotification {