+++ backend/canisters/community/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove bot API Keys ([#8131](https://github.com/open-chat-labs/open-chat/pull/8131))
 - Support bot making inline replies ([#8146](https://github.com/open-chat-labs/open-chat/pull/8146))
 - Pass in channelIds in the init args ([#8156](https://github.com/open-chat-labs/open-chat/pull/8156))
+- Add `initiated_by` to bot event notifications ([#8160](https://github.com/open-chat-labs/open-chat/pull/8160))
 
 ### Fixed